move SnapshotHolders to a separate file #35486
