+++ ydb/core/tx/columnshard/engines/snapshot_holders.h
@@ -1,0 +1,58 @@
+#pragma once
+
+#include "portions/portion_info.h"
+
+#include <algorithm>
+
+namespace NKikimr::NOlap {
+
+class TSnapshotHolders {
+    const TSnapshot MinReadSnapshot;
+    // Sorted from older to younger.
+    const std::vector<TSnapshot> TxInFlight;
+
+public:
+    TSnapshotHolders(TSnapshot minReadSnapshot, std::vector<TSnapshot> txInFlight)
+        : MinReadSnapshot(std::move(minReadSnapshot))
+        , TxInFlight(std::move(txInFlight)) {
+        AFL_VERIFY(std::is_sorted(TxInFlight.begin(), TxInFlight.end()));
+        if (!TxInFlight.empty()) {
+            AFL_VERIFY(TxInFlight.back() < MinReadSnapshot);
+        }
+    }
+
+    TSnapshot GetMinReadSnapshot() const {
+        return MinReadSnapshot;
+    }
+
+    bool CouldUsePortion(const TPortionInfo::TConstPtr& portion) const {
+        return CouldUse(
+            [&portion](const TSnapshot& snapshot) { return portion->IsRemovedFor(snapshot); },
+            [&portion](const TSnapshot& snapshot) { return portion->IsVisible(snapshot, true); });
+    }
+
+    template <class TIsRemovedFor, class TIsVisible>
+    bool CouldUse(const TIsRemovedFor& isRemovedFor, const TIsVisible& isVisible) const {
+        // The object can be used by new scans.
+        if (!isRemovedFor(MinReadSnapshot)) {
+            return true;
+        }
+
+        // Loop invariant: all txs older than txSnapshot couldn't use the object.
+        for (const auto& txSnapshot : TxInFlight) {
+            // This and all younger txs cannot see it.
+            if (isRemovedFor(txSnapshot)) {
+                return false;
+            }
+            // This tx could use it.
+            if (isVisible(txSnapshot)) {
+                return true;
+            }
+            // Current tx could not use it, maybe the next tx could.
+        }
+        // We have not found txs that could use it.
+        return false;
+    }
+};
+
+}   // namespace NKikimr::NOlap
